+++ ThirdPartyAdapters/mopub/mopub/src/main/java/com/mopub/mobileads/dfp/adapters/MoPubAdapter.java
@@ -38,6 +38,7 @@ import com.mopub.nativeads.ViewBinder;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.EnumSet;
@@ -284,7 +285,14 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
 
         String adunit = bundle.getString(MOPUB_AD_UNIT_KEY);
 
-        mAdSize = adSize;
+        mAdSize = getSupportedAdSize(context, adSize);
+        if (mAdSize == null) {
+          Log.w(TAG, "Failed to request ad, AdSize is null.");
+            if (mediationBannerListener != null) {
+              mediationBannerListener.onAdFailedToLoad(this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+            }
+          return;
+        }
         mMoPubView = new MoPubView(context);
         mMoPubView.setBannerAdListener(new MBannerListener(mediationBannerListener));
         mMoPubView.setAdUnitId(adunit);
@@ -308,6 +316,78 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
             initializeMoPub(context, adunit, mMoPubView, null, null);
         }
     }
+
+    AdSize getSupportedAdSize(Context context, AdSize adSize) {
+        AdSize original = new AdSize(adSize.getWidth(),
+            adSize.getHeight());
+
+        ArrayList<AdSize> potentials = new ArrayList<AdSize>(2);
+        potentials.add(AdSize.BANNER);
+        potentials.add(AdSize.MEDIUM_RECTANGLE);
+        potentials.add(AdSize.LEADERBOARD);
+        potentials.add(AdSize.WIDE_SKYSCRAPER);
+        Log.i(TAG, potentials.toString());
+        return findClosestSize(context, original, potentials);
+    }
+
+    // Start of helper code to remove when available in SDK
+    /**
+     * Find the closest supported AdSize from the list of potentials to the provided size.
+     * Returns null if none are within given threshold size range.
+     */
+    public static AdSize findClosestSize(
+        Context context, AdSize original, ArrayList<AdSize> potentials) {
+       if (potentials == null || original == null) {
+           return null;
+       }
+       float density = context.getResources().getDisplayMetrics().density;
+       int actualWidth = Math.round(original.getWidthInPixels(context)/density);
+       int actualHeight = Math.round(original.getHeightInPixels(context)/density);
+       original = new AdSize(actualWidth, actualHeight);
+        AdSize largestPotential = null;
+        for (AdSize potential : potentials) {
+            if (isSizeInRange(original, potential)) {
+                if (largestPotential == null) {
+                  largestPotential = potential;
+                } else {
+                  largestPotential = getLargerByArea(largestPotential, potential);
+                }
+            }
+        }
+        return largestPotential;
+    }
+
+    private static boolean isSizeInRange(AdSize original, AdSize potential) {
+        if (potential == null) {
+          return false;
+        }
+        double minWidthRatio = 0.5;
+        double minHeightRatio = 0.7;
+
+        int originalWidth = original.getWidth();
+        int potentialWidth = potential.getWidth();
+        int originalHeight = original.getHeight();
+        int potentialHeight = potential.getHeight();
+
+        if (originalWidth * minWidthRatio > potentialWidth ||
+            originalWidth < potentialWidth) {
+            return false;
+        }
+
+        if (originalHeight * minHeightRatio > potentialHeight ||
+            originalHeight < potentialHeight) {
+            return false;
+        }
+        return true;
+    }
+
+    private static AdSize getLargerByArea(AdSize size1, AdSize size2) {
+        int area1 = size1.getWidth() * size1.getHeight();
+        int area2 = size2.getWidth() * size2.getHeight();
+        return area1 > area2 ? size1 : size2;
+    }
+    // End code to remove when available in SDK
+
 
     @Override
     public View getBannerView() {
@@ -585,9 +665,9 @@ public class MoPubAdapter implements MediationNativeAdapter, MediationBannerAdap
          * @return a {@link Bundle} containing the specified extras.
          */
         public Bundle build() {
-            Bundle bundle = new Bundle();
-            bundle.putInt(ARG_PRIVACY_ICON_SIZE_DP, mPrivacyIconSizeDp);
-            return bundle;
+          Bundle bundle = new Bundle();
+          bundle.putInt(ARG_PRIVACY_ICON_SIZE_DP, mPrivacyIconSizeDp);
+          return bundle;
         }
     }
 }